refactor: model 패키지 하위에 dto, vo 패키지 위치
