+++ src/main/java/site/iris/issuefy/controller/AuthenticationController.java
@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import site.iris.issuefy.dto.OauthResponse;
+import site.iris.issuefy.response.OauthResponse;
 import site.iris.issuefy.entity.Jwt;
 import site.iris.issuefy.service.AuthenticationService;
 import site.iris.issuefy.service.TokenProvider;
-import site.iris.issuefy.vo.UserDto;
+import site.iris.issuefy.model.dto.UserDto;
 
 @Slf4j
 @RestController
