+++ src/main/java/site/iris/issuefy/model/dto/UserDto.java
@@ -1,4 +1,4 @@
-package site.iris.issuefy.vo;
+package site.iris.issuefy.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
