+++ src/main/java/site/iris/issuefy/model/vo/OrgRecord.java
@@ -1,6 +1,8 @@
-package site.iris.issuefy.vo;
+package site.iris.issuefy.model.vo;
 
 import java.util.List;
+
+import site.iris.issuefy.model.dto.RepositoryDto;
 
 public record OrgRecord(Long id, String name, List<RepositoryDto> repositories) {
     public static OrgRecord from(Long id, String name, List<RepositoryDto> repositories) {
