+++ src/main/java/site/iris/issuefy/response/SubscribeResponse.java
@@ -1,6 +1,6 @@
-package site.iris.issuefy.dto;
+package site.iris.issuefy.response;
 
-import site.iris.issuefy.vo.OrgRecord;
+import site.iris.issuefy.model.vo.OrgRecord;
 
 public record SubscribeResponse(OrgRecord org) {
 	public static SubscribeResponse from(OrgRecord org) {
