+++ src/main/java/site/iris/issuefy/service/SubscribeService.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import site.iris.issuefy.dto.SubscribeResponse;
+import site.iris.issuefy.response.SubscribeResponse;
 import site.iris.issuefy.entity.Org;
 import site.iris.issuefy.entity.Repository;
 import site.iris.issuefy.entity.Subscribe;
@@ -19,10 +19,9 @@ import site.iris.issuefy.repository.OrgRepository;
 import site.iris.issuefy.repository.RepositoryRepository;
 import site.iris.issuefy.repository.SubscribeRepository;
 import site.iris.issuefy.repository.UserRepository;
-import site.iris.issuefy.vo.OrgRecord;
-import site.iris.issuefy.vo.RepositoryDto;
-import site.iris.issuefy.vo.RepositoryUrlDto;
-import site.iris.issuefy.vo.TokenDto;
+import site.iris.issuefy.model.vo.OrgRecord;
+import site.iris.issuefy.model.dto.RepositoryDto;
+import site.iris.issuefy.model.dto.RepositoryUrlDto;
 
 @Service
 @RequiredArgsConstructor
