+++ src/main/java/site/iris/issuefy/service/UserService.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 import site.iris.issuefy.entity.User;
 import site.iris.issuefy.repository.UserRepository;
-import site.iris.issuefy.vo.UserDto;
-import site.iris.issuefy.vo.UserVerifyDto;
+import site.iris.issuefy.model.dto.UserDto;
+import site.iris.issuefy.model.dto.UserVerifyDto;
 
 @Service
 @RequiredArgsConstructor
